requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.5
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.5
django-adminlte3==0.1.6
django-betterforms==1.2
django-cloudinary-storage==0.3.0
django-extensions==3.1.0
django-heroku==0.3.1
django-phonenumber-field==5.0.0
gunicorn==20.0.4
idna==2.10
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.8.1
pexpect==4.8.0
phonenumbers==8.12.18
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
ptyprocess==0.7.0
pycodestyle==2.6.0
pydotplus==2.0.2
Pygments==2.8.0
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
pyzmq==22.0.3
requests==2.25.1
rope==0.18.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
